支持导入数据时指定表名

+++ src/commands/LoadCommand.php
@@ -1,6 +1,7 @@
 <?php
 namespace winwin\db\tools\commands;
 
+use kuiper\helper\Arrays;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputOption;
@@ -17,6 +18,7 @@ class LoadCommand extends BaseCommand
         parent::configure();
         $this->setName("load")
             ->setDescription("Load data to database table")
+            ->addOption('--table', null, InputOption::VALUE_REQUIRED, "Table to load")
             ->addOption('--format', '-f', InputOption::VALUE_REQUIRED, "Input data format, support json|yaml|php")
             ->addOption('--truncate', '-t', InputOption::VALUE_NONE, "Truncate table before load data")
             ->addArgument('file', InputArgument::OPTIONAL, "Data input file, default read from stdin");
@@ -27,6 +29,7 @@ class LoadCommand extends BaseCommand
         $truncate = $input->getOption('truncate');
         $format = $input->getOption('format');
         $file = $input->getArgument('file');
+        $table = $input->getOption('table');
         if (empty($file)) {
             $file = "php://stdin";
             if (empty($format)) {
@@ -34,25 +37,32 @@ class LoadCommand extends BaseCommand
             }
         }
         $dataset = DataDumper::loadFile($file, $format);
+        if ($table) {
+            $dataset = [$table => $dataset];
+        }
         $db = $this->getConnection($input);
         try {
             foreach ($dataset as $table => $rows) {
+                if (empty($rows)) {
+                    continue;
+                }
+                $columns = $db->getSchemaManager()->listTableColumns($table);
+                $fields = array_intersect(array_keys($rows[0]), Arrays::pull($columns, 'name', Arrays::GETTER));
                 if ($truncate) {
                     $db->executeUpdate("truncate `$table`");
                 }
                 $batchSize = 1000;
                 foreach (array_chunk($rows, $batchSize) as $batchRows) {
-                    $row = $batchRows[0];
                     $sql = sprintf(
-                        "INSERT INTO `%s` (`%s`) VALUES",
+                        'INSERT INTO `%s` (`%s`) VALUES',
                         $table,
-                        implode("`,`", array_keys($row))
+                        implode('`,`', $fields)
                     );
-                    $sql .= implode(",", array_fill(0, count($batchRows), sprintf('(%s)', implode(",", array_fill(0, count($row), "?")))));
-                    $keys = array_keys($row);
+                    $rowPlaceholder = sprintf('(%s)', implode(',', array_fill(0, count($fields), '?')));
+                    $sql .= implode(',', array_fill(0, count($batchRows), $rowPlaceholder));
                     $bindValues = [];
                     foreach ($batchRows as $row) {
-                        foreach ($keys as $name) {
+                        foreach ($fields as $name) {
                             $bindValues[] = isset($row[$name]) ? $row[$name] : null;
                         }
                     }
